Fix building on Windows
