+++ internal/log.go
@@ -18,7 +18,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"log/syslog"
 	"net/http"
 	"os"
 	"path"
@@ -31,7 +30,6 @@ import (
 	"github.com/matrix-org/dendrite/setup/config"
 	"github.com/matrix-org/dugong"
 	"github.com/sirupsen/logrus"
-	lSyslog "github.com/sirupsen/logrus/hooks/syslog"
 )
 
 type utcFormatter struct {
@@ -108,37 +106,6 @@ func SetupStdLogging() {
 	})
 }
 
-// SetupHookLogging configures the logging hooks defined in the configuration.
-// If something fails here it means that the logging was improperly configured,
-// so we just exit with the error
-func SetupHookLogging(hooks []config.LogrusHook, componentName string) {
-	logrus.SetReportCaller(true)
-	for _, hook := range hooks {
-		// Check we received a proper logging level
-		level, err := logrus.ParseLevel(hook.Level)
-		if err != nil {
-			logrus.Fatalf("Unrecognised logging level %s: %q", hook.Level, err)
-		}
-
-		// Perform a first filter on the logs according to the lowest level of all
-		// (Eg: If we have hook for info and above, prevent logrus from processing debug logs)
-		if logrus.GetLevel() < level {
-			logrus.SetLevel(level)
-		}
-
-		switch hook.Type {
-		case "file":
-			checkFileHookParams(hook.Params)
-			setupFileHook(hook, level, componentName)
-		case "syslog":
-			checkSyslogHookParams(hook.Params)
-			setupSyslogHook(hook, level, componentName)
-		default:
-			logrus.Fatalf("Unrecognised logging hook type: %s", hook.Type)
-		}
-	}
-}
-
 // File type hooks should be provided a path to a directory to store log files
 func checkFileHookParams(params map[string]interface{}) {
 	path, ok := params["path"]
@@ -176,34 +143,6 @@ func setupFileHook(hook config.LogrusHook, level logrus.Level, componentName str
 			&dugong.DailyRotationSchedule{GZip: true},
 		),
 	})
-}
-
-func checkSyslogHookParams(params map[string]interface{}) {
-	addr, ok := params["address"]
-	if !ok {
-		logrus.Fatalf("Expecting a parameter \"address\" for logging hook of type \"syslog\"")
-	}
-
-	if _, ok := addr.(string); !ok {
-		logrus.Fatalf("Parameter \"address\" for logging hook of type \"syslog\" should be a string")
-	}
-
-	proto, ok2 := params["protocol"]
-	if !ok2 {
-		logrus.Fatalf("Expecting a parameter \"protocol\" for logging hook of type \"syslog\"")
-	}
-
-	if _, ok2 := proto.(string); !ok2 {
-		logrus.Fatalf("Parameter \"protocol\" for logging hook of type \"syslog\" should be a string")
-	}
-
-}
-
-func setupSyslogHook(hook config.LogrusHook, level logrus.Level, componentName string) {
-	syslogHook, err := lSyslog.NewSyslogHook(hook.Params["protocol"].(string), hook.Params["address"].(string), syslog.LOG_INFO, componentName)
-	if err == nil {
-		logrus.AddHook(&logLevelHook{level, syslogHook})
-	}
 }
 
 //CloseAndLogIfError Closes io.Closer and logs the error if any
